+++ src/helpers/string.ts
@@ -4,11 +4,9 @@ export function isSpecialSymbol(s: string, i: number): boolean {
   return c < 65 || (c > 90 && (c < 97 || (c > 123 && c < 128)));
 }
 
-/**
- * Returns count of chars in lower case (for any language with ignoring numbers, symbols)
+/** Returns count of chars in lower case (for any language with ignoring numbers, symbols)
  * @param s
- * @param stopWith point to calc up to this value (for performance reason)
- */
+ * @param stopWith point to calc up to this value (for performance reason) */
 export function stringLowerCount(s: string, stopWith?: number): number {
   let c = 0;
   stopWith = stopWith ?? s.length;
@@ -20,11 +18,9 @@ export function stringLowerCount(s: string, stopWith?: number): number {
   return c;
 }
 
-/**
- * Returns count of chars in upper case (for any language with ignoring numbers)
+/** Returns count of chars in upper case (for any language with ignoring numbers)
  * @param s
- * @param stopWith point to calc up to this value (for performance reason)
- */
+ * @param stopWith point to calc up to this value (for performance reason) */
 export function stringUpperCount(s: string, stopWith?: number): number {
   let c = 0;
   stopWith = stopWith ?? s.length;
@@ -37,22 +33,49 @@ export function stringUpperCount(s: string, stopWith?: number): number {
 }
 
 /**
- * Changes camelCase 'somePropValue' to 'Some Prop Value'
+ ** Changes camelCase 'somePropValue' to 'Some Prop Value'
+ ** Changes snake_case 'some_prop_value' to 'Some Prop Value'
+ ** Changes kebab-case 'some-prop-value' to 'Some Prop Value';
  *
- * Changes snakeCase 'some_prop_value' to 'Some prop value'
- *
- * Changes kebabCase 'some-prop-value' to 'Some prop value'; `false` by default
- *
- * To capitalize use css [text-transform: capitalize]
  * @param text The string to change
- * @param changeKebabCase Set true to apply kebabCase rule; `false` by default
- * @returns Prettified string
- */
-export function stringPrettify(text: string, changeKebabCase = false): string {
-  const r = text
-    .replace(/([A-ZА-Я])/g, " $1")
-    .trimStart()
-    .replace(new RegExp(`[_${(changeKebabCase && "-") || ""}]`, "g"), " ")
-    .replace(/[ ]{2,}/, " ");
-  return r.charAt(0).toUpperCase() + r.slice(1);
+ * @param capitalize Set uppercase 1st letter of every word: `somePropValue` => `Some Prop Value`; @defaultValue `true`
+ * @returns Prettified string */
+export function stringPrettify(text: string, capitalize = true, handleKebabCase = false): string {
+  let r = "";
+  let nextUpper = false; // mostly it means next word
+  let wasUpper = false; // tracks if the previous character was uppercase
+  for (let i = 0; i < text.length; ++i) {
+    const c = text.charCodeAt(i);
+    if (c > 96 && c < 123) {
+      // Lowercase: a-z
+      wasUpper = false;
+      r += String.fromCharCode(nextUpper || i === 0 ? c - 32 : c); // snake_case & kebab_case
+    } else if (c > 64 && c < 91) {
+      // Uppercase: A-Z
+      const cNext = text.charCodeAt(i + 1);
+      const isAbbr = i === text.length - 1 || (cNext > 64 && cNext < 91); // nextChar is abbreviation
+      const isAbbrPrev = wasUpper && isAbbr; // prevChar + cur + next is abbreviation
+      // eslint-disable-next-line prefer-template
+      if (!nextUpper && !isAbbrPrev && i !== 0) {
+        r += " ";
+      }
+      r += String.fromCharCode(capitalize || isAbbr || i === 0 ? c : c + 32); // camelCase
+      wasUpper = true;
+    } else if (c === 95 || (c === 45 && handleKebabCase)) {
+      wasUpper = false;
+      // Underscore (_) or dash (-)
+      /* istanbul ignore else */
+      if (i !== 0) {
+        r += " ";
+      }
+      nextUpper = capitalize;
+      continue;
+    } else {
+      r += String.fromCharCode(c);
+      wasUpper = false;
+    }
+    nextUpper = false;
+  }
+
+  return r;
 }