+++ test/jest/helpers/helper.stringPrettify.test.js
@@ -4,21 +4,29 @@ describe("helper.stringPrettify", () => {
   test("camelCase", () => {
     expect(stringPrettify("somePropValue")).toEqual("Some Prop Value");
     expect(stringPrettify("TestMe")).toEqual("Test Me");
-  });
-  test("russian-camelCase", () => {
-    expect(stringPrettify("хорошоСделано")).toEqual("Хорошо Сделано");
+    expect(stringPrettify("abcAzpqZpq")).toEqual("Abc Azpq Zpq");
+    expect(stringPrettify("abcAzpqZpq", false)).toEqual("Abc azpq zpq");
+
+    expect(stringPrettify("SMSReminder")).toEqual("SMS Reminder");
+    expect(stringPrettify("ReminderSMS")).toEqual("Reminder SMS");
+
+    expect(stringPrettify("SMSReminder", false)).toEqual("SMS reminder");
+    expect(stringPrettify("reminderSMS")).toEqual("Reminder SMS");
+    expect(stringPrettify("reminderSMS", false)).toEqual("Reminder SMS");
   });
   test("snakeCase", () => {
-    expect(stringPrettify("some_prop_value")).toEqual("Some prop value");
-    expect(stringPrettify("some_Prop_value")).toEqual("Some Prop value");
+    expect(stringPrettify("some_prop_value")).toEqual("Some Prop Value");
+    expect(stringPrettify("some_Prop_value")).toEqual("Some Prop Value");
+    expect(stringPrettify("some_prop_value", false)).toEqual("Some prop value");
   });
   test("kebabCase", () => {
     expect(stringPrettify("some-prop-value")).toEqual("Some-prop-value");
-    expect(stringPrettify("some-prop-value", true)).toEqual("Some prop value");
-    expect(stringPrettify("some-prop-Value", true)).toEqual("Some prop Value");
+    expect(stringPrettify("some-prop-value", true, true)).toEqual("Some Prop Value");
+    expect(stringPrettify("some-prop-value", false, true)).toEqual("Some prop value");
   });
   test("mixCase", () => {
-    expect(stringPrettify("some_prop-valueHere")).toEqual("Some prop-value Here");
-    expect(stringPrettify("some_prop-valueHere", true)).toEqual("Some prop value Here");
+    expect(stringPrettify("some_prop-valueHere")).toEqual("Some Prop-value Here");
+    expect(stringPrettify("some_prop-valueHere", false)).toEqual("Some prop-value here");
+    expect(stringPrettify("some_prop-valueHere", true, true)).toEqual("Some Prop Value Here");
   });
 });